backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.0
drf-yasg==1.21.7
